Move field descriptors to their own file
